refactor: folder the token list height into a single hook
