+++ packages/widget/src/pages/SelectTokenPage/SelectTokenPage.tsx
@@ -1,30 +1,26 @@
 import { Box } from '@mui/material';
 import type { FC } from 'react';
-import { useLayoutEffect, useRef, useState } from 'react';
+import { useRef } from 'react';
 import { useTranslation } from 'react-i18next';
 import { ChainSelect } from '../../components/ChainSelect/ChainSelect.js';
 import { PageContainer } from '../../components/PageContainer.js';
 import { TokenList } from '../../components/TokenList/TokenList.js';
-import { useContentHeight } from '../../hooks/useContentHeight.js';
 import { useHeader } from '../../hooks/useHeader.js';
 import { useNavigateBack } from '../../hooks/useNavigateBack.js';
 import { useScrollableOverflowHidden } from '../../hooks/useScrollableContainer.js';
 import { useSwapOnly } from '../../hooks/useSwapOnly.js';
+import { useTokenListHeight } from '../../hooks/useTokenListHeight.js';
 import { useWidgetConfig } from '../../providers/WidgetProvider/WidgetProvider.js';
 import type { FormTypeProps } from '../../stores/form/types.js';
 import { SearchTokenInput } from './SearchTokenInput.js';
-
-const minTokenListHeight = 360;
 
 export const SelectTokenPage: FC<FormTypeProps> = ({ formType }) => {
   useScrollableOverflowHidden();
   const { navigateBack } = useNavigateBack();
   const headerRef = useRef<HTMLElement>(null);
   const listParentRef = useRef<HTMLUListElement | null>(null);
-  const contentHeight = useContentHeight({ listParentRef });
+  const tokenListHeight = useTokenListHeight({ listParentRef, headerRef });
 
-  // TODO: question: can we move this in to useContentHeightHook
-  const [tokenListHeight, setTokenListHeight] = useState(0);
   const swapOnly = useSwapOnly();
 
   const { subvariant } = useWidgetConfig();
@@ -37,16 +33,6 @@ export const SelectTokenPage: FC<FormTypeProps> = ({ formType }) => {
       : t(`header.to`);
 
   useHeader(title);
-
-  // TODO: question: can we move this in to useContentHeightHook
-  useLayoutEffect(() => {
-    setTokenListHeight(
-      Math.max(
-        contentHeight - (headerRef.current?.offsetHeight ?? 0),
-        minTokenListHeight,
-      ),
-    );
-  }, [contentHeight]);
 
   const hideChainSelect = swapOnly && formType === 'to';
