+++ packages/widget/src/pages/TransactionPage/StatusBottomSheet.tsx
@@ -12,9 +12,9 @@ import { BottomSheet } from '../../components/BottomSheet/BottomSheet.js';
 import type { BottomSheetBase } from '../../components/BottomSheet/types.js';
 import { Token } from '../../components/Token/Token.js';
 import { useAvailableChains } from '../../hooks/useAvailableChains.js';
-import { useSetContentHeight } from '../../hooks/useContentHeight.js';
 import { useNavigateBack } from '../../hooks/useNavigateBack.js';
 import { getProcessMessage } from '../../hooks/useProcessMessage.js';
+import { useSetContentHeight } from '../../hooks/useSetContentHeight.js';
 import { useTokenBalance } from '../../hooks/useTokenBalance.js';
 import { useWidgetConfig } from '../../providers/WidgetProvider/WidgetProvider.js';
 import { useFieldActions } from '../../stores/form/useFieldActions.js';
