+++ packages/widget/src/pages/TransactionPage/TokenValueBottomSheet.tsx
@@ -6,7 +6,7 @@ import { forwardRef, useRef } from 'react';
 import { useTranslation } from 'react-i18next';
 import { BottomSheet } from '../../components/BottomSheet/BottomSheet.js';
 import type { BottomSheetBase } from '../../components/BottomSheet/types.js';
-import { useSetContentHeight } from '../../hooks/useContentHeight.js';
+import { useSetContentHeight } from '../../hooks/useSetContentHeight.js';
 import { CenterContainer, IconCircle } from './StatusBottomSheet.style.js';
 import { calcValueLoss } from './utils.js';
